+++ src/attachment_loader.rs
@@ -4,20 +4,12 @@ use crate::{
         spAttachmentLoader_dispose,
     },
     c_interface::{NewFromPtr, SyncPtr},
-    Atlas, Attachment, AttachmentType, RegionProps, Skin,
+    Atlas, Attachment, AttachmentType, RegionProps, Skin, SpineError,
 };
 
-/// Error types related to [`AttachmentLoader`](`crate::AttachmentLoader`).
-#[derive(Debug)]
-pub enum AttachmentLoaderError {
-    /// Creating an attachment failed.
-    /// Check [`error1`](`Self::error1`) and [`error2`](`Self::error2`) for more information.
-    CreateAttachmentFailed,
-    InvalidArgument {
-        field: &'static str,
-    },
-}
-
+/// A loader for creating custom attachments.
+///
+/// Currently only supports [`Atlas`](`crate::Atlas`) based attachments.
 #[derive(Debug)]
 pub struct AttachmentLoader {
     c_attachment_loader: SyncPtr<spAttachmentLoader>,
@@ -46,20 +38,18 @@ impl AttachmentLoader {
     ///
     /// # Errors
     ///
-    /// Returns [`AttachmentLoaderError::CreateAttachmentFailed`] if creating the attachment failed.
+    /// Returns [`SpineError::CreationFailed`] if creating the attachment failed.
     /// Check [`error1`](`Self::error1`) and [`error2`](`Self::error2`) for more information.
-    /// Returns [`AttachmentLoaderError::InvalidArgument`] if `name` or `path` contain a null byte.
+    /// Returns [`SpineError::NulError`] if `name` or `path` contain a null byte.
     pub fn create_attachment(
         &self,
         skin: Option<Skin>,
         attachment_type: AttachmentType,
         name: &str,
         path: &str,
-    ) -> Result<Attachment, AttachmentLoaderError> {
-        let c_name = std::ffi::CString::new(name)
-            .map_err(|_| AttachmentLoaderError::InvalidArgument { field: "name" })?;
-        let c_path = std::ffi::CString::new(path)
-            .map_err(|_| AttachmentLoaderError::InvalidArgument { field: "path" })?;
+    ) -> Result<Attachment, SpineError> {
+        let c_name = std::ffi::CString::new(name)?;
+        let c_path = std::ffi::CString::new(path)?;
 
         unsafe {
             let c_name = c_name.as_ptr();
@@ -77,7 +67,7 @@ impl AttachmentLoader {
             );
 
             if attachment.is_null() {
-                Err(AttachmentLoaderError::CreateAttachmentFailed)
+                Err(SpineError::new_creation_failed("Attachment"))
             } else {
                 Ok(Attachment::new_from_ptr(attachment))
             }
@@ -88,20 +78,20 @@ impl AttachmentLoader {
     ///
     /// # Errors
     ///
-    /// Returns [`AttachmentLoaderError::CreateAttachmentFailed`] if creating the attachment failed.
+    /// Returns [`SpineError::CreationFailed`] if creating the attachment failed.
     /// Check [`error1`](`Self::error1`) and [`error2`](`Self::error2`) for more information.
-    /// Returns [`AttachmentLoaderError::InvalidArgument`] if `name` or `path` contain a null byte.
+    /// Returns [`SpineError::NulError`] if `name` or `path` contain a null byte.
     pub fn create_region_attachment(
         &self,
         skin: Option<Skin>,
         name: &str,
         path: &str,
         props: &RegionProps,
-    ) -> Result<Attachment, AttachmentLoaderError> {
+    ) -> Result<Attachment, SpineError> {
         let attachment = self.create_attachment(skin, AttachmentType::Region, name, path)?;
 
         let Some(mut region) = attachment.as_region() else {
-            return Err(AttachmentLoaderError::CreateAttachmentFailed);
+            return Err(SpineError::new_creation_failed("RegionAttachment"));
         };
 
         region.update_from_props(props);
@@ -112,12 +102,6 @@ impl AttachmentLoader {
     c_accessor_string!(error1, error1);
     c_accessor_string!(error2, error2);
     c_ptr!(c_attachment_loader, spAttachmentLoader);
-}
-
-impl Clone for AttachmentLoader {
-    fn clone(&self) -> Self {
-        unsafe { AttachmentLoader::new_from_ptr(self.c_ptr()) }
-    }
 }
 
 impl Drop for AttachmentLoader {